feat: robot got moves
