+++ src/Types.tsx
@@ -1,0 +1,14 @@
+export enum FACES {
+    North = 'NORTH',
+    South = 'SOUTH',
+    East = 'EAST',
+    West = 'WEST',
+}
+
+export enum COMMANDS {
+    Place = 'PLACE',
+    Move = 'MOVE',
+    Left = 'LEFT',
+    Right = 'RIGHT',
+    Report = 'REPORT',
+}
